feat(typescript): Migrate all the components to typescript
