+++ .babelrc
@@ -1,11 +1,11 @@
-{
-  "presets": ["@babel/preset-react", "@babel/preset-env"],
-  "plugins": [
-    [
-      "module-resolver",
-      {
-        "root": ["./src"]
-      }
-    ]
-  ]
-}
+// {
+//   "presets": ["@babel/preset-react", "@babel/preset-env"],
+//   "plugins": [
+//     [
+//       "module-resolver",
+//       {
+//         "root": ["./src"]
+//       }
+//     ]
+//   ]
+// }
