+++ package.json
@@ -23,46 +23,53 @@
     "formik": "^2.1.5",
     "formik-material-ui": "^3.0.0-alpha.0",
     "prop-types": "^15.7.2",
+    "react": "16.13.1",
     "react-dom": "^16.13.1",
     "react-spinners": "^0.9.0",
-    "yup": "^0.29.3",
-    "react": "16.13.1"
+    "yup": "^0.29.3"
   },
   "devDependencies": {
     "@babel/cli": "^7.10.5",
     "@babel/core": "^7.11.1",
     "@babel/preset-env": "^7.11.0",
     "@babel/preset-react": "^7.10.4",
+    "@material-ui/core": "^4.11.0",
+    "@material-ui/icons": "^4.9.1",
     "@semantic-release/changelog": "^5.0.1",
     "@semantic-release/commit-analyzer": "^8.0.1",
     "@semantic-release/git": "^9.0.0",
     "@semantic-release/github": "^7.0.7",
     "@semantic-release/npm": "^7.0.5",
     "@semantic-release/release-notes-generator": "^9.0.1",
+    "@types/node": "^14.6.0",
+    "@types/react": "^16.9.47",
+    "@types/react-dom": "^16.9.8",
     "babel-loader": "^8.1.0",
     "babel-plugin-module-resolver": "^4.0.0",
     "css-loader": "^4.2.1",
-    "file-loader": "^6.0.0",
-    "sass-loader": "^9.0.3",
-    "semantic-release": "^17.0.8",
-    "style-loader": "^1.2.1",
-    "webpack-cli": "^3.3.12",
-    "react-scripts": "3.4.2",
-    "@material-ui/core": "^4.11.0",
-    "@material-ui/icons": "^4.9.1",
     "env-cmd": "^10.1.0",
+    "file-loader": "^6.0.0",
     "formik": "^2.1.5",
     "formik-material-ui": "^3.0.0-alpha.0",
     "prop-types": "^15.7.2",
+    "react": "16.13.1",
+    "react-docgen-typescript-loader": "^3.7.2",
     "react-dom": "^16.13.1",
+    "react-scripts": "3.4.2",
     "react-spinners": "^0.9.0",
-    "yup": "^0.29.3",
-    "react": "16.13.1"
+    "sass-loader": "^9.0.3",
+    "semantic-release": "^17.0.8",
+    "style-loader": "^1.2.1",
+    "ts-loader": "^8.0.3",
+    "typescript": "^4.0.2",
+    "webpack-cli": "^3.3.12",
+    "yup": "^0.29.3"
   },
   "scripts": {
     "build": "webpack --mode production",
     "test": "react-scripts test --passWithNoTests",
-    "semantic-release": "npx semantic-release"
+    "semantic-release": "npx semantic-release",
+    "yalc:link": "yarn build && yalc publish"
   },
   "keywords": [
     "components",
