+++ src/components/DynamicForm/DynamicForm.tsx
@@ -14,11 +14,10 @@ import {
   Typography,
 } from '@material-ui/core';
 import { BounceLoader } from 'react-spinners';
-import grasTheme from './../../styles/theme';
-import { makeStyles, withStyles } from '@material-ui/core/styles';
+import grasTheme from '../../styles/theme';
+import { makeStyles } from '@material-ui/core/styles';
 import * as Yup from 'yup';
-import PropTypes from 'prop-types';
-// a
+
 // const fields = [
 //   {
 //     name: 'email',
@@ -75,43 +74,44 @@ import PropTypes from 'prop-types';
 //     ],
 //   },
 // ];
-
-const contactStyle = {
-  color: '#00A4EC',
-  textAlign: 'center',
+type DynamicFormProps = {
+  disableButtonOnLoad: boolean;
+  fields: Array<object>;
+  validationSchema: any;
+  onSubmitForm: Function;
+  title: string;
 };
 
-const dynamicFormStyle = makeStyles({
-  h4: {
+const useStyles = makeStyles({
+  mainTitle: {
     fontSize: 24,
     fontWeight: 700,
-    color: '#0B4566',
+    color: '#00A4EC',
+    textAlign: 'center',
   },
 });
-const DynamicFormComponent = ({
+
+const DynamicForm = ({
   disableButtonOnLoad = true,
   fields = [],
   validationSchema = Yup.object(),
   onSubmitForm = () => {},
   title = '',
-}) => {
+}: DynamicFormProps) => {
+  const classes = useStyles();
   return (
     <ThemeProvider theme={grasTheme}>
       <Paper>
         <Box p={4}>
           {title && (
             <Box pb={2}>
-              <Typography
-                variant="h4"
-                className={dynamicFormStyle.h4}
-                style={contactStyle}
-              >
+              <Typography variant="h4" className={classes.mainTitle}>
                 {title}
               </Typography>
             </Box>
           )}
           <Formik
-            initialValues={fields.reduce((accumulator, element) => {
+            initialValues={fields.reduce((accumulator, element: any = {}) => {
               accumulator = {
                 ...accumulator,
                 [element.name]: element.initialValue,
@@ -138,7 +138,7 @@ const DynamicFormComponent = ({
                 )}
                 {!isSubmitting && (
                   <Grid container>
-                    {fields.map((item, i) => {
+                    {fields.map((item: any = {}, i) => {
                       if (item.componentType === 'TextField') {
                         return (
                           <Grid
@@ -235,14 +235,4 @@ const DynamicFormComponent = ({
   );
 };
 
-DynamicFormComponent.propTypes = {
-  title: PropTypes.string,
-  disableButtonOnLoad: PropTypes.bool,
-  validationSchema: PropTypes.object,
-  fields: PropTypes.array,
-  onSubmitForm: PropTypes.func,
-};
-const DynamicForm = withStyles(dynamicFormStyle, { withTheme: true })(
-  DynamicFormComponent
-);
-export { DynamicForm };
+export default DynamicForm;
