--- src/components/DynamicForm/index.js
@@ -1,1 +1,0 @@
-export { DynamicForm } from './DynamicForm';
