+++ src/components/Information/partials/About/About.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import * as React from 'react';
 import { Box, Link, Typography } from '@material-ui/core';
 
 const About = () => {
