+++ src/components/Information/partials/About/index.ts
@@ -1,0 +1,1 @@
+export { default } from './About';
