+++ src/components/Information/partials/Disclaimer/Disclaimer.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import * as React from 'react';
 import { Box, Typography } from '@material-ui/core';
 
 const Disclaimer = () => {
