+++ src/components/Information/partials/Disclaimer/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Disclaimer';
