--- src/components/Information/partials/FAQ/index.js
@@ -1,1 +1,0 @@
-export { default } from './FAQ.js';
