+++ src/components/Information/partials/FAQ/index.ts
@@ -1,0 +1,1 @@
+export { default } from './FAQ';
