+++ src/components/Information/partials/Privacy/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Privacy';
