--- src/components/Information/partials/Privacy/styles.js
@@ -1,5 +1,0 @@
-export default {
-  link: {
-    color: 'blue',
-  },
-};
