+++ src/components/Information/partials/Terms/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Terms';
