+++ src/components/Information/partials/index.ts
@@ -1,0 +1,6 @@
+export { default as About } from './About';
+export { default as Contact } from './Contact';
+export { default as Disclaimer } from './Disclaimer';
+export { default as FAQ } from './FAQ';
+export { default as Privacy } from './Privacy';
+export { default as Terms } from './Terms';
