+++ src/components/SidebarTitle/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SidebarTitle';
