--- src/index.js
@@ -1,5 +1,0 @@
-import { DynamicForm } from './components/DynamicForm';
-import { Information } from './components/Information';
-import { SidebarTitle } from './components/Title';
-
-export { Information, DynamicForm, SidebarTitle };
