+++ src/index.ts
@@ -1,0 +1,4 @@
+import { default as DynamicForm } from './components/DynamicForm';
+import { default as SidebarTitle } from './components/SidebarTitle';
+import { default as Information } from './components/Information';
+export { SidebarTitle, DynamicForm, Information };
