+++ tsconfig.json
@@ -1,0 +1,13 @@
+{
+  "include": ["src/*"],
+  "compilerOptions": {
+    "module": "commonjs",
+    "target": "es5",
+    "jsx": "react",
+    "allowSyntheticDefaultImports": true,
+    "lib": ["es2015", "dom"],
+    "outDir": "dist",
+    "sourceMap": true,
+    "declaration": true
+  }
+}
